Move football::CPlayTime traits
